feat: extension内核新增端口桥接能力与权限规范化构建器

+++ backend/internal/extension/kernel/permission/canonical_builder.go
@@ -95,7 +95,7 @@ func BuildEvaluationRequestFromInvocation(subject PermissionSubject, requirement
 		WithIsBackground(inv.IsBackground).
 		WithScopeSnapshotID(inv.ScopeSnapshotID).
 		WithApprovalMode(string(inv.ApprovalMode)).
-		WithGeneration(int(inv.Generation)).
+		WithGeneration(int64(inv.Generation)).
 		WithExecutionContext(ExecutionContextFromInvocation(inv)).
 		Build()
 }
